Collapse golf/f1 into unified sport array + clippy cleanups
Golf and F1 now live as variants of `SportSection` under the `sport`
config key, discriminated by the inner `"sport"` field. A single config
array can mix team sports, golf, and F1 — matching how they share the
same conceptual slot.

Also folds in clippy fixes: derivable Default impls, removed unnecessary
borrows in filelib, and replaced `&input == ""` with `is_empty()`.

+++ src/lib/matrix/sport.rs
@@ -190,8 +190,8 @@ impl SportMatrix {
 
         // Names slide left at one-pixel-per-tick. The slot is 22px wide; once a
         // name fully scrolls off, reset by modulo so it loops.
-        let home_x = -((xpos as i32) % (home_name.chars().count() as i32 * 4 + 22).max(1));
-        let away_x = 44 - ((xpos as i32) % (away_name.chars().count() as i32 * 4 + 22).max(1));
+        let home_x = -(xpos % (home_name.chars().count() as i32 * 4 + 22).max(1));
+        let away_x = 44 - (xpos % (away_name.chars().count() as i32 * 4 + 22).max(1));
 
         draw_text(img, font, home_x, baseline, Color::WHITE, home_name);
         draw_text(img, font, 28, baseline, Color::WHITE, "VS");
